relocate viewcount
